requirements.txt
sense-hat==2.4.0
pyzmq==25.1.0
protobuf==4.30.*
